REF: Create StorageExtensionDtype
